requirements.txt
requests
python-dotenv

pytrends
beautifulsoup4
fpdf2==2.7.6
